requirements.txt
PySide6
praw
openai
python-dotenv